// src/state/types.ts
export type SerializedBigNumber = string

interface SerializedFarmUserData {
  allowance: string
  tokenBalance: string
  stakedBalance: string
  earnings: string
  canHarvest: boolean
}
// ...
export interface SerializedStake extends SerializedStakeConfig {
  tokenPriceBusd?: string
  quoteTokenPriceBusd?: string
  tokenAmountTotal?: SerializedBigNumber
  lpTotalInQuoteToken?: SerializedBigNumber
  lpTotalSupply?: SerializedBigNumber
  tokenPriceVsQuote?: SerializedBigNumber
  poolWeight?: SerializedBigNumber
  userData?: SerializedFarmUserData
}
// ...
export interface SerializedStakesState {
  data: SerializedStake[]
  loadArchivedStakesData: boolean
  userDataLoaded: boolean
}